Use float for dsda_DegreesToAngle

+++ prboom2/src/dsda/utility.c
@@ -301,7 +301,7 @@ int dsda_IntToFixed(int x)
 }
 
 // ANG1 is off by 256 / 360 due to rounding
-angle_t dsda_DegreesToAngle(int x)
+angle_t dsda_DegreesToAngle(float x)
 {
   return ANG1 * x + 256 * x / 360;
 }
